# contracts/market-data/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "vibe-invest Financial Data",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/financial-context": {
      "post": {
        "summary": "Financial Context",
        "operationId": "createFinancialContext",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/financial-overview": {
      "post": {
        "summary": "Financial Overview",
        "operationId": "getFinancialOverview",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinancialOverviewResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/financial-metric-series": {
      "post": {
        "summary": "Financial Metric Series Endpoint",
        "operationId": "getFinancialMetricSeries",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "metric",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Metric"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedFactResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/valuation-evidence": {
      "post": {
        "summary": "Valuation Evidence Endpoint",
        "operationId": "getValuationEvidence",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValuationEvidenceResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/filing-document": {
      "post": {
        "summary": "Filing Document",
        "operationId": "readFilingDocument",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "filing_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filing Id"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FilingDocumentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/news-search": {
      "post": {
        "summary": "News Search",
        "operationId": "searchNews",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Keyword"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactQueryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/news-document": {
      "post": {
        "summary": "News Document",
        "operationId": "readNewsDocument",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsDocumentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewsDocumentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/web-search": {
      "post": {
        "summary": "Web Search",
        "operationId": "searchWebEvidence",
        "parameters": [
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Query"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactQueryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/company-events": {
      "post": {
        "summary": "Company Events",
        "operationId": "listCompanyEvents",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactQueryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/official-company-events": {
      "post": {
        "summary": "Official Company Events",
        "operationId": "listOfficialCompanyEvents",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactQueryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/technical-indicators": {
      "post": {
        "summary": "Technical Indicators",
        "operationId": "getTechnicalIndicators",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FactQueryResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/technical-evidence": {
      "post": {
        "summary": "Technical Evidence",
        "operationId": "getTechnicalEvidence",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechnicalEvidenceResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/price-window": {
      "post": {
        "summary": "Price Window",
        "operationId": "getPriceWindow",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Symbol"
            }
          },
          {
            "name": "start_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Start Date"
            }
          },
          {
            "name": "end_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "End Date"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 60,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriceWindowResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/quotes": {
      "post": {
        "summary": "Quotes",
        "operationId": "createQuoteBatch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "type": "string"
                },
                "type": "array",
                "title": "Symbols"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuoteBatch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "FilingDocumentResult": {
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/AtomicFact"
            },
            "type": "array",
            "title": "Facts"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "eligibility": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Eligibility"
          },
          "returnedCount": {
            "type": "integer",
            "title": "Returnedcount"
          },
          "totalCount": {
            "type": "integer",
            "title": "Totalcount"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          },
          "items": {
            "items": {},
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "facts",
          "returnedCount",
          "totalCount",
          "truncated",
          "items"
        ],
        "title": "FilingDocumentResult"
      },
      "FinancialContext": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "title": "Fetched At"
          },
          "quote": {
            "$ref": "#/components/schemas/CapabilityResult"
          },
          "history": {
            "$ref": "#/components/schemas/CapabilityResult"
          },
          "news": {
            "$ref": "#/components/schemas/CapabilityResult"
          },
          "fundamentals": {
            "$ref": "#/components/schemas/CapabilityResult"
          },
          "indicators": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Indicators"
              },
              {
                "type": "null"
              }
            ]
          },
          "valuation": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Valuation"
          },
          "valuation_sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "type": "array",
            "title": "Valuation Sources",
            "default": []
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/AtomicFact"
            },
            "type": "array",
            "title": "Facts"
          },
          "gaps": {
            "items": {
              "$ref": "#/components/schemas/DataGap"
            },
            "type": "array",
            "title": "Gaps"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "fetched_at",
          "quote",
          "history",
          "news",
          "fundamentals",
          "indicators",
          "facts",
          "gaps"
        ],
        "title": "FinancialContext"
      },
      "FinancialOverviewResult": {
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/AtomicFact"
            },
            "type": "array",
            "title": "Facts"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "eligibility": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Eligibility"
          },
          "overview": {
            "title": "Overview"
          }
        },
        "type": "object",
        "required": [
          "facts",
          "overview"
        ],
        "title": "FinancialOverviewResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NewsDocumentRequest": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/AtomicFact"
          }
        },
        "type": "object",
        "required": [
          "candidate"
        ],
        "title": "NewsDocumentRequest"
      },
...
      "PaginatedFactResult": {
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/AtomicFact"
            },
            "type": "array",
            "title": "Facts"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "eligibility": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Eligibility"
          },
          "returnedCount": {
            "type": "integer",
            "title": "Returnedcount"
          },
          "totalCount": {
            "type": "integer",
            "title": "Totalcount"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          }
        },
        "type": "object",
        "required": [
          "facts",
          "returnedCount",
          "totalCount",
          "truncated"
        ],
        "title": "PaginatedFactResult"
      },
...
      "ValuationEvidenceResult": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "authorizedComparables": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Authorizedcomparables"
          },
          "comparables": {
            "items": {
              "$ref": "#/components/schemas/ValuationComparable"
            },
            "type": "array",
            "title": "Comparables"
          },
          "currentMultiples": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Currentmultiples"
          },
          "historicalRanges": {
            "additionalProperties": {
              "items": {
                "type": "number"
              },
              "type": "array"
            },
            "type": "object",
            "title": "Historicalranges"
          },
          "methods": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ValuationMethodView"
            },
            "type": "object",
            "title": "Methods"
          },
          "facts": {
            "items": {
              "$ref": "#/components/schemas/AtomicFact"
            },
            "type": "array",
            "title": "Facts"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceStatus"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "authorizedComparables",
          "comparables",
          "currentMultiples",
          "historicalRanges",
          "methods",
          "facts"
        ],
        "title": "ValuationEvidenceResult"
      },
...
    }
  }
}